refactor: child - rename package, Controller, Service
