+++ isol/server-child/src/main/java/shinhan/server_child/domain/user/dto/PhoneFindRequest.java
@@ -1,4 +1,4 @@
-package shinhan.server_child.domain.child.dto;
+package shinhan.server_child.domain.user.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -9,12 +9,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class LoginInfoFindRequest {
+public class PhoneFindRequest {
 
     @NotBlank(message = "전화번호를 입력해주세요.")
     @Pattern(regexp = "\\d{3}-\\d{4}-\\d{4}", message = "전화번호 형식이 올바르지 않습니다.")
     private String phoneNum;
-    @NotBlank(message = "비밀번호를 입력해주세요.")
-    @Pattern(regexp = "\\d{6}", message = "비밀번호는 6자리 숫자여야 합니다.")
-    private String accountInfo;
 }
